Add __typename support
https://github.com/Bartosz-D3V/grafik/issues/71

* Extract data structures to separate package.
