+++ client/helper.go
@@ -1,3 +1,4 @@
+// Package client contains the code used internally by grafik to prepare & send HTTP requests.
 package client
 
 import "strings"
