+++ ds/helper.go
@@ -1,0 +1,8 @@
+// Package ds (Data Structure) contains all golang data structures used by generator.
+package ds
+
+// isPrimitive checks if passed golang type is primitive.
+// It only checks golang primitives generated by evaluator.
+func isPrimitive(s string) bool {
+	return s == "string" || s == "int" || s == "bool"
+}
