+++ generator/const.go
@@ -1,3 +1,4 @@
+// Package generator abstracts writing specific Go constructs (interfaces, structs etc) into IO.
 package generator
 
 // Header - default header added to the generated grafik client.
